+++ lib/core/resources/manager_assets.dart
@@ -2,6 +2,8 @@ class ManagerAssets {
   static const String azkarImage = 'assets/images/azkar.jpg';
   static const String celebrationImage = 'assets/images/celebration.png';
 
+  static const String defaultProfileImage = 'assets/images/default_profile_image.jpg';
+
   static const String sliderImage1 = 'assets/images/slider/slider1.png';
 
   static const String productImage = 'assets/images/products/product-1.png';
