+++ lib/core/resources/manager_colors.dart
@@ -6,6 +6,7 @@ class ManagerColors {
   static Color primaryColor = const Color(0xFF957255);
   static Color secondaryColor = Colors.brown;
   static Color textColor = const Color(0xFF6E6E6F);
+  static Color secondaryTextColor = const Color(0xFF83848A);
   static Color borderColor = const Color(0xFFE8E8E8);
   static Color iconsColor = const Color(0xFF909090);
   static Color drawerBackgroundColor = const Color(0xFF474340);
@@ -14,6 +15,7 @@ class ManagerColors {
   static Color discountColor = const Color(0xFFCE2621);
   static Color newProductColor = const Color(0xFF00AE3B);
   static Color yellow = const Color(0xFFFFF500);
+  static Color textButtonUnderLineColor = const Color(0xFFCDCDCD);
   static Color brown200 = Colors.brown.shade300;
   static Color transparent = Colors.transparent;
   static Color white70 = Colors.white70;
