+++ lib/core/resources/manager_height.dart
@@ -5,13 +5,16 @@ class ManagerHeights {
   static const double h8 = 8;
   static const double h10 = 10;
   static const double h12 = 12;
+  static const double h14 = 14;
   static const double h16 = 16;
   static const double h24 = 24;
   static const double h20 = 20;
+  static const double h28 = 28;
   static const double h30 = 30;
   static const double h35 = 35;
   static const double h36 = 36;
   static const double h40 = 40;
+  static const double h48 = 48;
   static const double h50 = 50;
   static const double h60 = 60;
   static const double h65 = 65;
