+++ lib/core/resources/manager_margin.dart
@@ -4,6 +4,7 @@ class ManagerMargin {
   static const double v8 = 8;
   static const double v12 = 12;
   static const double v24 = 24;
+  static const double v30 = 30;
   static const double v48 = 48;
 
   static const double h8 = 8;
