--- lib/core/resources/manager_strings.dart
@@ -1,6 +1,0 @@
-class ManagerStrings {
-  static const String appName = 'أذكار';
-  static const String aboutScreen = 'About Screen';
-
-  static const String hintTextUserName = 'ادخل اسمك';
-}
