+++ lib/core/resources/manager_width.dart
@@ -8,6 +8,7 @@ class ManagerWidth {
   static const double w16 = 16;
   static const double w20 = 20;
   static const double w25 = 25;
+  static const double w28 = 28;
   static const double w30 = 30;
   static const double w35 = 35;
   static const double w36 = 36;
