+++ lib/core/widgets/base_text_widget.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:store_app/core/resources/manager_font_weight.dart';
+import 'package:store_app/core/resources/manager_fonts.dart';
 
 import '../resources/manager_colors.dart';
 import '../resources/manager_font_sizes.dart';
@@ -18,7 +19,7 @@ Text baseText(
     style: TextStyle(
       fontSize: fontSize ?? ManagerFontSizes.s16,
       color: color ?? ManagerColors.textColor,
-      fontFamily: fontFamily ?? '',
+      fontFamily: fontFamily ?? ManagerFont.appFont,
       fontWeight: fontWeight ?? ManagerFontWeight.w400,
     ),
   );
